rename skeleton to auth service #3
