+++ backend/Dockerfile.auth
@@ -1,0 +1,10 @@
+FROM gradle:5.5-jdk8 as builder
+WORKDIR /builder
+ADD . /builder
+RUN gradle -p auth build --stacktrace
+
+FROM openjdk:8-jre-alpine
+WORKDIR /app
+COPY --from=builder /builder/auth/build/libs/auth-0.0.1-SNAPSHOT.jar .
+CMD ["java", "-jar", "auth-0.0.1-SNAPSHOT.jar"]
+EXPOSE 8080
