+++ backend/auth/build.gradle.kts
@@ -21,11 +21,11 @@ val ktlint by configurations.creating
 dependencies {
 	ktlint("com.github.shyiko:ktlint:0.31.0")
 
-	implementation("org.springframework.boot:spring-boot-starter-web")
-	implementation("com.fasterxml.jackson.module:jackson-module-kotlin")
 	implementation("org.jetbrains.kotlin:kotlin-reflect")
 	implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk8")
-	testImplementation("org.springframework.boot:spring-boot-starter-test")
+	implementation("org.springframework.boot:spring-boot-starter-web:2.1.7.RELEASE")
+	testImplementation("org.springframework.boot:spring-boot-starter-test:2.1.7.RELEASE")
+	implementation("com.fasterxml.jackson.module:jackson-module-kotlin:2.9.9")
 }
 
 tasks.withType<KotlinCompile> {
