+++ backend/auth/src/main/kotlin/com/oursky/presentation/microservices/kotlin/auth/controller/HealthcheckController.kt
@@ -1,0 +1,13 @@
+package com.oursky.presentation.microservices.kotlin.auth.controller
+
+import org.springframework.web.bind.annotation.RestController
+import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.http.ResponseEntity
+
+@RestController
+public class HealthcheckController {
+    @GetMapping("/healthcheck")
+    fun healthcheck(): ResponseEntity<Void> {
+        return ResponseEntity.ok().build()
+    }
+}
