+++ backend/docker-compose.yml
@@ -1,9 +1,9 @@
 version: '3'
 services:
-  backend:
+  auth:
     image: gradle:5.5-jdk8
     working_dir: /app
-    command: ./gradlew bootRun
+    command: ./gradlew --no-daemon -p auth bootRun
     ports:
       - 8080:8080
     volumes:
